Refactor: 코드 가독성 향상을 위한 띄움 및 캐러셀 내용 상수처리
