+++ src/components/common/Button.tsx
@@ -3,7 +3,6 @@ import { forwardRef } from 'react';
 
 import { cn } from '@lib/cn';
 
-// Figma 디자인에 맞는 버튼 스타일 정의
 const buttonVariants = cva(
   'inline-flex items-center justify-center w-auto whitespace-nowrap rounded-md font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50',
   {
