+++ src/router/ProtectedRoute.tsx
@@ -10,6 +10,7 @@ interface ProtectedRouteProps {
 
 const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
   const isAuthenticated = useAuthStore((state) => state.isAuthenticated);
+  
   const location = useLocation();
 
   if (!isAuthenticated) {
